+++ docs/changelog.md
@@ -5,6 +5,7 @@ Unreleased
 - Remove the `autoInstallMissingIgs` configuration parameter, it is now enabled when `httpReadOnly` is not [#315](https://github.com/ahdis/matchbox/issues/315)
 - Improve the Matchbox server documentation [#315](https://github.com/ahdis/matchbox/issues/315)
 - Respect the 'onlyOneEngine' mode in MappingLanguageInterceptor [#316](https://github.com/ahdis/matchbox/issues/316)
+- Use the proper encoding when returning a transformed resource [#318](https://github.com/ahdis/matchbox/issues/318)
 
 2024/11/25 Release 3.9.8